Refactor: Mount Data-Management Handlers as a Child Resource
Registration had no way to express "this handler is not part of the
public API". Every public method became a route in one flat namespace,
and the only lever for hiding one was a leading underscore, which also
lies about a method's Python visibility — setup_schema is called from
__init__ and from tests and could not be renamed.

api/admin_resource.py takes the eleven data-management handlers, the
eleven helpers only they reach, prefer_score_tuner, and the module-level
constants and imports none of the rest of the file used. APIResource
mounts it under a path prefix via the same build_route_table its own
registration uses, so the child cannot end up registered by different
rules than its parent.

The seam is five methods and four handles, derived from the AST rather
than guessed: _reload_engine, _run_query, _serve_static_file,
_set_statement_timeout and _setup_complete, plus _conn_pool,
_cache_generation, _last_import_time and _setup_complete_cache. The child
holds a parent reference for those. That is a routing boundary, not a
decoupling, and is written down as such.

advertise=False is passed once at the mount rather than repeated on each
handler, so forgetting it is a property of the single call site. The 404
listing filters on the flag, which keeps it from becoming a directory of
what is behind the mount — the failure mode that would leave every test
passing.

get_migrations loses its route; db_utils.get_migrations stays for
in-process use. Nothing called it over HTTP.

24 tests in test_admin_mount.py, including the public route table as a
literal set. Eight bare hasattr smoke checks are deleted rather than
ported — the literal supersedes them. scripts/bench_search.py's raw
attribute assignment now targets the child; it is outside the test suite,
so override_attr would not have caught it.

api_resource.py: 2307 -> 1306 lines.
