+++ api/tests/test_cubecobra_and_prefer_scores.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
 
 def _insert_card(api: APIResource, raw: dict) -> uuid.UUID:
     """Insert a raw card and return its oracle_id as a UUID."""
-    api._upsert_cards([raw])
+    api.admin._upsert_cards([raw])
     (processed,) = preprocess_card(raw)
     return uuid.UUID(processed["oracle_id"])
 
@@ -34,7 +34,7 @@ class TestInsertCubecobraData:
         oracle_id = _insert_card(api_resource, make_raw_card(name="Cubecobra Insert Test"))
 
         cubecobra_data = {oracle_id: {"elo": 1200.5, "cube_count": 42, "pick_count": 100}}
-        rows_updated = api_resource._insert_cubecobra_data(cubecobra_data)
+        rows_updated = api_resource.admin._insert_cubecobra_data(cubecobra_data)
 
         assert rows_updated >= 1
 
@@ -52,11 +52,11 @@ class TestInsertCubecobraData:
 
     def test_unknown_oracle_id_updates_zero_rows(self, api_resource: APIResource) -> None:
         unknown = uuid.uuid4()
-        rows_updated = api_resource._insert_cubecobra_data({unknown: {"elo": 999.0, "cube_count": 1, "pick_count": 1}})
+        rows_updated = api_resource.admin._insert_cubecobra_data({unknown: {"elo": 999.0, "cube_count": 1, "pick_count": 1}})
         assert rows_updated == 0
 
     def test_empty_dict_updates_zero_rows(self, api_resource: APIResource) -> None:
-        rows_updated = api_resource._insert_cubecobra_data({})
+        rows_updated = api_resource.admin._insert_cubecobra_data({})
         assert rows_updated == 0
 
     def test_multiple_cards_updated_in_one_call(self, api_resource: APIResource) -> None:
@@ -67,7 +67,7 @@ class TestInsertCubecobraData:
             oid1: {"elo": 1100.0, "cube_count": 10, "pick_count": 20},
             oid2: {"elo": 900.0, "cube_count": 5, "pick_count": 8},
         }
-        rows_updated = api_resource._insert_cubecobra_data(cubecobra_data)
+        rows_updated = api_resource.admin._insert_cubecobra_data(cubecobra_data)
         assert rows_updated == 2
 
 
@@ -86,12 +86,12 @@ class TestFetchCubecobraData:
         oracle_id = uuid.uuid4()
         page1 = [{"oracle_id": str(oracle_id), "elo": 1500, "cubeCount": 30, "pickCount": 60}]
 
-        with patch.object(api_resource, "_session") as mock_session, patch("api.api_resource.time.sleep"):
+        with patch.object(api_resource.admin, "_session") as mock_session, patch("api.api_resource.time.sleep"):
             mock_session.get.side_effect = [
                 self._mock_response(page1),
                 self._mock_response([]),  # empty page terminates
             ]
-            pages = list(api_resource._fetch_cubecobra_data({oracle_id}))
+            pages = list(api_resource.admin._fetch_cubecobra_data({oracle_id}))
 
         assert len(pages) == 1
         assert oracle_id in pages[0]
@@ -105,9 +105,9 @@ class TestFetchCubecobraData:
             {"oracle_id": str(unknown), "elo": 800, "cubeCount": 2, "pickCount": 4},
         ]
 
-        with patch.object(api_resource, "_session") as mock_session, patch("api.api_resource.time.sleep"):
+        with patch.object(api_resource.admin, "_session") as mock_session, patch("api.api_resource.time.sleep"):
             mock_session.get.side_effect = [self._mock_response(page1), self._mock_response([])]
-            pages = list(api_resource._fetch_cubecobra_data({known}))
+            pages = list(api_resource.admin._fetch_cubecobra_data({known}))
 
         assert known in pages[0]
         assert unknown not in pages[0]
@@ -121,18 +121,18 @@ class TestFetchCubecobraData:
             [],  # terminator
         ]
 
-        with patch.object(api_resource, "_session") as mock_session, patch("api.api_resource.time.sleep"):
+        with patch.object(api_resource.admin, "_session") as mock_session, patch("api.api_resource.time.sleep"):
             mock_session.get.side_effect = [self._mock_response(p) for p in pages_data]
-            pages = list(api_resource._fetch_cubecobra_data(set(oids)))
+            pages = list(api_resource.admin._fetch_cubecobra_data(set(oids)))
 
         assert len(pages) == 3
 
     def test_empty_db_oracle_ids_yields_empty_pages(self, api_resource: APIResource) -> None:
         page1 = [{"oracle_id": str(uuid.uuid4()), "elo": 1, "cubeCount": 1, "pickCount": 1}]
 
-        with patch.object(api_resource, "_session") as mock_session, patch("api.api_resource.time.sleep"):
+        with patch.object(api_resource.admin, "_session") as mock_session, patch("api.api_resource.time.sleep"):
             mock_session.get.side_effect = [self._mock_response(page1), self._mock_response([])]
-            pages = list(api_resource._fetch_cubecobra_data(set()))
+            pages = list(api_resource.admin._fetch_cubecobra_data(set()))
 
         # All cards filtered out, but we still get one page dict (empty)
         assert all(len(p) == 0 for p in pages)
@@ -145,21 +145,21 @@ class TestFetchCubecobraData:
 
 class TestBackfillPreferScores:
     def test_returns_success_status(self, api_resource: APIResource) -> None:
-        result = api_resource.backfill_prefer_scores()
+        result = api_resource.admin.backfill_prefer_scores()
         assert result["status"] == "success"
 
     def test_returns_cards_updated_count(self, api_resource: APIResource) -> None:
         _insert_card(api_resource, make_raw_card(name=f"Prefer Score Card {uuid.uuid4()}"))
-        result = api_resource.backfill_prefer_scores()
+        result = api_resource.admin.backfill_prefer_scores()
         assert result["cards_updated"] >= 1
 
     def test_message_includes_count(self, api_resource: APIResource) -> None:
-        result = api_resource.backfill_prefer_scores()
+        result = api_resource.admin.backfill_prefer_scores()
         assert str(result["cards_updated"]) in result["message"]
 
     def test_prefer_score_populated_in_db(self, api_resource: APIResource) -> None:
         oracle_id = _insert_card(api_resource, make_raw_card(name=f"Prefer Score Check {uuid.uuid4()}"))
-        api_resource.backfill_prefer_scores()
+        api_resource.admin.backfill_prefer_scores()
 
         with api_resource._conn_pool.connection() as conn, conn.cursor() as cursor:
             cursor.execute("SELECT prefer_score FROM magic.cards WHERE oracle_id = %s LIMIT 1", (oracle_id,))
@@ -171,18 +171,18 @@ class TestBackfillPreferScores:
     def test_second_run_updates_zero_rows(self, api_resource: APIResource) -> None:
         """Re-running the backfill on already-scored cards should touch no rows."""
         _insert_card(api_resource, make_raw_card(name=f"Idempotent Score Card {uuid.uuid4()}"))
-        api_resource.backfill_prefer_scores()
+        api_resource.admin.backfill_prefer_scores()
 
-        result = api_resource.backfill_prefer_scores()
+        result = api_resource.admin.backfill_prefer_scores()
 
         assert result["cards_updated"] == 0
 
     def test_reports_duration_and_scored_count(self, api_resource: APIResource) -> None:
         """cards_scored counts the whole scored corpus, not just the rows this run moved."""
         _insert_card(api_resource, make_raw_card(name=f"Stats Card {uuid.uuid4()}"))
-        api_resource.backfill_prefer_scores()
+        api_resource.admin.backfill_prefer_scores()
 
-        result = api_resource.backfill_prefer_scores()
+        result = api_resource.admin.backfill_prefer_scores()
 
         assert result["duration_seconds"] >= 0
         # Nothing moved on the second run, but the corpus is still fully scored.
@@ -197,13 +197,13 @@ class TestBackfillPreferScores:
 
 class TestBackfillCubecobraScores:
     def test_returns_success_status(self, api_resource: APIResource) -> None:
-        result = api_resource.backfill_cubecobra_scores()
+        result = api_resource.admin.backfill_cubecobra_scores()
         assert result["status"] == "success"
 
     def test_reports_duration_and_counts(self, api_resource: APIResource) -> None:
         _insert_card(api_resource, make_raw_card(name=f"Cubecobra Stats Card {uuid.uuid4()}"))
 
-        result = api_resource.backfill_cubecobra_scores()
+        result = api_resource.admin.backfill_cubecobra_scores()
 
         assert result["duration_seconds"] >= 0
         assert result["cards_updated"] >= 0
@@ -216,8 +216,8 @@ class TestBackfillCubecobraScores:
         from one computed over an all-NULL corpus.
         """
         oracle_id = _insert_card(api_resource, make_raw_card(name=f"Cubecobra Data Card {uuid.uuid4()}"))
-        before = api_resource.backfill_cubecobra_scores()["cards_with_cubecobra_data"]
+        before = api_resource.admin.backfill_cubecobra_scores()["cards_with_cubecobra_data"]
 
-        api_resource._insert_cubecobra_data({oracle_id: {"elo": 1500.0, "cube_count": 7, "pick_count": 9}})
+        api_resource.admin._insert_cubecobra_data({oracle_id: {"elo": 1500.0, "cube_count": 7, "pick_count": 9}})
 
-        assert api_resource.backfill_cubecobra_scores()["cards_with_cubecobra_data"] == before + 1
+        assert api_resource.admin.backfill_cubecobra_scores()["cards_with_cubecobra_data"] == before + 1
